move export/import buttons to About page
