+++ lib/widgets/about_page.dart
@@ -19,6 +19,7 @@ import 'package:cuppa_mobile/common/globals.dart';
 import 'package:cuppa_mobile/common/icons.dart';
 import 'package:cuppa_mobile/common/padding.dart';
 import 'package:cuppa_mobile/common/text_styles.dart';
+import 'package:cuppa_mobile/data/export.dart';
 import 'package:cuppa_mobile/data/localization.dart';
 import 'package:cuppa_mobile/data/provider.dart';
 import 'package:cuppa_mobile/widgets/list_divider.dart';
@@ -102,6 +103,9 @@ class AboutWidget extends StatelessWidget {
                       child: listDivider,
                     ),
                   ),
+                  // Tools: export/import data
+                  _exportImportTools(context),
+                  listDivider,
                   // How to report issues
                   _listItem(
                     title: AppString.issues.translate(),
@@ -179,6 +183,155 @@ class AboutWidget extends StatelessWidget {
         contentPadding: listTilePadding,
         dense: true,
       ),
+    );
+  }
+
+  // Tools: export/import data
+  Widget _exportImportTools(BuildContext context) {
+    AppProvider provider = Provider.of<AppProvider>(context);
+
+    return Align(
+      alignment: Alignment.topLeft,
+      child: IgnorePointer(
+        // Disable export/import while timer is active
+        ignoring: provider.activeTeas.isNotEmpty,
+        child: Opacity(
+          opacity: provider.activeTeas.isNotEmpty ? fadeOpacity : noOpacity,
+          child: ListTile(
+            iconColor: Theme.of(context).colorScheme.onPrimaryContainer,
+            title: Text(
+              AppString.export_import.translate(),
+              style: textStyleTitle,
+            ),
+            trailing: Row(
+              mainAxisSize: MainAxisSize.min,
+              children: [
+                _exportButton(context),
+                const VerticalDivider(),
+                _importButton(context),
+              ],
+            ),
+            contentPadding: listTilePadding,
+            dense: true,
+          ),
+        ),
+      ),
+    );
+  }
+
+  // Export button
+  Widget _exportButton(BuildContext context) {
+    AppProvider provider = Provider.of<AppProvider>(context);
+
+    return IconButton(
+      icon: getPlatformExportIcon(),
+      onPressed: () {
+        // Attempt to save an export file and report if failed
+        Export.create(provider, share: true).then(
+          (exported) {
+            if (!exported) {
+              showAdaptiveDialog(
+                context: context,
+                barrierDismissible: false,
+                builder: (BuildContext context) {
+                  return AlertDialog.adaptive(
+                    content: SingleChildScrollView(
+                      child: ListBody(
+                        children: <Widget>[
+                          Text(AppString.export_failure.translate()),
+                        ],
+                      ),
+                    ),
+                    actions: [
+                      adaptiveDialogAction(
+                        isDefaultAction: true,
+                        text: AppString.ok_button.translate(),
+                        onPressed: () => Navigator.of(context).pop(false),
+                      ),
+                    ],
+                  );
+                },
+              );
+            }
+          },
+        );
+      },
+    );
+  }
+
+  // Import button
+  Widget _importButton(BuildContext context) {
+    AppProvider provider = Provider.of<AppProvider>(context);
+
+    return IconButton(
+      icon: getPlatformImportIcon(),
+      onPressed: () async {
+        bool confirmed = await _confirmImport(context);
+        if (confirmed) {
+          // Attempt to load an export file and report the result
+          Export.load(provider).then(
+            (imported) => showAdaptiveDialog(
+              context: context,
+              barrierDismissible: false,
+              builder: (BuildContext context) {
+                return AlertDialog.adaptive(
+                  content: SingleChildScrollView(
+                    child: ListBody(
+                      children: <Widget>[
+                        Text(
+                          imported
+                              ? AppString.import_sucess.translate()
+                              : AppString.import_failure.translate(),
+                        ),
+                      ],
+                    ),
+                  ),
+                  actions: [
+                    adaptiveDialogAction(
+                      isDefaultAction: true,
+                      text: AppString.ok_button.translate(),
+                      onPressed: () => Navigator.of(context).pop(false),
+                    ),
+                  ],
+                );
+              },
+            ),
+          );
+        }
+      },
+    );
+  }
+
+  // Import confirmation dialog
+  Future _confirmImport(BuildContext context) {
+    return showAdaptiveDialog(
+      context: context,
+      barrierDismissible: false,
+      builder: (BuildContext context) {
+        return AlertDialog.adaptive(
+          title: Text(AppString.confirm_title.translate()),
+          content: SingleChildScrollView(
+            child: ListBody(
+              children: <Widget>[
+                Text(AppString.confirm_import.translate()),
+                const SizedBox(height: 14.0),
+                Text(AppString.confirm_continue.translate()),
+              ],
+            ),
+          ),
+          actions: [
+            adaptiveDialogAction(
+              isDefaultAction: true,
+              text: AppString.no_button.translate(),
+              onPressed: () => Navigator.of(context).pop(false),
+            ),
+            adaptiveDialogAction(
+              text: AppString.yes_button.translate(),
+              onPressed: () => Navigator.of(context).pop(true),
+            ),
+          ],
+        );
+      },
     );
   }
